Refactor temperature handling in GUI components to use float instead of int for improved precision. Updated temperature display formats to show one decimal place. Added new methods in TempInput for handling float temperatures.

+++ src/slic3r/GUI/DeviceCore/DevExtruderSystem.cpp
@@ -74,7 +74,7 @@ namespace Slic3r
     {
         if (extder_id < 0 || extder_id >= static_cast<int>(m_extders.size()))
             return;
-        m_extders[extder_id].set_target_temp(temp);
+        m_extders[extder_id].set_target_temp(static_cast<float>(temp));
     }
 
     bool DevExtderSystem::CanQuitSwitching() const
